Refactor and solve some issues
