+++ src/modules/lessons/lessons.router.js
@@ -6,6 +6,8 @@ import AddStudentToLessonsModule from './lesson-students'
 import AddMaterialsToLessonsModule from './lesson-material-list'
 import ListOfMaterialsForLessonsModule from './materials-for-lesson-list'
 import StudentLessons from './student-lessons'
+import ManageStudentsMaterialsModule from './student-materials'
+import ManageStudentsRecordingsModule from './student-recordings'
 
 /**
  * STUDENT ROUTES
@@ -15,15 +17,38 @@ const studentRoute = [
   {
     path: '/',
     component: Module,
-    redirect: { name: 'manage-lessons' },
+    redirect: { name: 'my-lessons' },
     children: [
       {
         path: 'lessons',
-        name: 'my-lessons',
-        component: StudentLessons,
-        meta: {
-          middleware: [authMiddleware]
-        }
+        component: Module,
+        redirect: { name: 'my-lessons' },
+        children: [
+          {
+            path: 'list',
+            name: 'my-lessons',
+            component: StudentLessons,
+            meta: {
+              middleware: [authMiddleware]
+            }
+          },
+          {
+            path: 'materials',
+            name: 'manage-students-materials',
+            component: ManageStudentsMaterialsModule,
+            meta: {
+              middleware: [authMiddleware]
+            }
+          },
+          {
+            path: 'recordings',
+            name: 'manage-students-recordings',
+            component: ManageStudentsRecordingsModule,
+            meta: {
+              middleware: [authMiddleware]
+            }
+          }
+        ]
       }
     ]
   }
