+++ src/modules/students/router.js
@@ -1,9 +1,6 @@
 import Module from './module.vue'
 import authMiddleware from '@/middlewares/auth'
 import ManageStudentsModule from '@/modules/students/views/Students'
-import ManageStudentsLessonsModule from '@/modules/students/views/StudentsLessons'
-import ManageStudentsMaterialsModule from '@/modules/students/views/StudentsMaterials/index.vue'
-import ManageStudentsRecordingsModule from '@/modules/students/views/StudentsRecordings/index.vue'
 
 const shortcutRoutes = [
   {
@@ -30,30 +27,6 @@ const moduleRoute = [
             path: 'list',
             name: 'manage-students',
             component: ManageStudentsModule,
-            meta: {
-              middleware: [authMiddleware]
-            }
-          },
-          {
-            path: 'lessons',
-            name: 'manage-students-lessons',
-            component: ManageStudentsLessonsModule,
-            meta: {
-              middleware: [authMiddleware]
-            }
-          },
-          {
-            path: 'materials',
-            name: 'manage-students-materials',
-            component: ManageStudentsMaterialsModule,
-            meta: {
-              middleware: [authMiddleware]
-            }
-          },
-          {
-            path: 'recordings',
-            name: 'manage-students-recordings',
-            component: ManageStudentsRecordingsModule,
             meta: {
               middleware: [authMiddleware]
             }
